added the data layer and defined the dependency injection.

+++ app/src/main/java/com/retailcloudandroid/retailcloudmachinetestandroid/domain/entity/CartItem.kt
@@ -4,10 +4,10 @@ data class CartItem(
     val itemId: String,
     val itemName: String,
     val sellingPrice: Double,
-    val taxPercentage: Int,
+    val taxPercentage: Double,
     val quantity: Int
 ) {
     val subtotal: Double = sellingPrice * quantity
-    val taxAmount: Double = subtotal * (taxPercentage.toDouble() / 100.0)
+    val taxAmount: Double = subtotal * (taxPercentage / 100.0)
     val total: Double = subtotal + taxAmount
 }
