+++ app/src/main/java/com/retailcloudandroid/retailcloudmachinetestandroid/domain/entity/Item.kt
@@ -4,5 +4,5 @@ data class Item(
     val id: String,
     val name: String,
     val sellingPrice: Double,
-    val taxPercentage: Int
+    val taxPercentage: Double
 )
